+++ packages/backend-graphql/src/factory.test.ts
@@ -18,6 +18,8 @@ import { Role, ServicePrincipal } from 'aws-cdk-lib/aws-iam';
 import {
   CfnIdentityPool,
   CfnIdentityPoolRoleAttachment,
+  CfnUserPool,
+  CfnUserPoolClient,
   UserPool,
   UserPoolClient,
 } from 'aws-cdk-lib/aws-cognito';
@@ -85,6 +87,10 @@ void describe('DataFactory', () => {
             assumedBy: new ServicePrincipal('test.amazon.com'),
           }),
           cfnResources: {
+            userPool: new CfnUserPool(stack, 'CfnUserPool', {}),
+            userPoolClient: new CfnUserPoolClient(stack, 'CfnUserPoolClient', {
+              userPoolId: 'userPool',
+            }),
             identityPool: new CfnIdentityPool(stack, 'identityPool', {
               allowUnauthenticatedIdentities: true,
             }),